refactor user wip#2

+++ backend/apps/users/src/app/user-info/user-info.repository.ts
@@ -4,66 +4,70 @@ import { CRUDRepository } from '@backend/util/util-types';
 import { adaptPrismaUser } from '../utils/adapt-prisma-user';
 import { UserQuery } from '@backend/shared-quieries';
 import { FitnessLevel, UserRole } from '@prisma/client';
-import { UserInfo } from '@backend/shared/shared-types';
-import { UserInfoEntity } from './entity/user-info-entity.type';
+import { User } from '@backend/shared/shared-types';
+import { UserInfoEntity } from './user-info.entity';
 
 @Injectable()
 export class UserInfoRepository
-  implements CRUDRepository<UserInfoEntity, number, UserInfo>
+  implements CRUDRepository<UserInfoEntity, number, User>
 {
   constructor(private readonly prisma: PrismaService) {}
 
-  public async create(item: UserInfoEntity): Promise<UserInfo> {
-  const data ={
-    ...item.toObject(),
-    password:item.passwordHash
-  }
-  delete data._id;
-  delete data.passwordHash;
-  const newUser = await this.prisma.user.create({ data } );
-  return adaptPrismaUser(newUser);
-  }
+  // public async create(item: UserInfoEntity): Promise<UserInfo> {
+  // const data ={
+  //   ...item.toObject(),
+  //   password:item.passwordHash
+  // }
+  // delete data._id;
+  // delete data.passwordHash;
+  // const newUser = await this.prisma.user.create({ data } );
+  // return adaptPrismaUser(newUser);
+  // }
 
-  public async findById(id: number): Promise<UserInfo | null> {
+  public async findById(id: number): Promise<User | null> {
     const user = await this.prisma.user.findFirst({
       where: {
-        id,
+        userId: id,
+      },
+      include: {
+        sportsmanInfo: true,
+        coachInfo: true,
       },
     });
     return adaptPrismaUser(user);
   }
 
-  public async findByEmail(email: string): Promise<UserInfo | null> {
-    const user = await this.prisma.user.findFirst({
-      where: {
-        email,
-      },
-    });
-    return adaptPrismaUser(user);
-  }
+  // public async findByEmail(email: string): Promise<UserInfo | null> {
+  //   const user = await this.prisma.user.findFirst({
+  //     where: {
+  //       email,
+  //     },
+  //   });
+  //   return adaptPrismaUser(user);
+  // }
 
-  public async show({limit, page, role, location, fitnessLevel, workoutType}:UserQuery): Promise<UserInfo[]> {
-    const queryParams = {
-      where: {
-        AND: {
-          role: role as UserRole,
-          fitnessLevel: fitnessLevel as FitnessLevel,
-          workoutType: undefined,
-          location: undefined,
-        }
-      },
-      take: limit,
-      skip: page > 0 ? limit * (page - 1) : undefined,
-    }
-    if (workoutType) {
-      queryParams.where.AND.workoutType = { hasSome: workoutType };
-    }
-    if (location) {
-      queryParams.where.AND.location = { in: location };
-    }
-    const users = await this.prisma.user.findMany(queryParams);
-    return users.map((user)=> adaptPrismaUser(user));
-  }
+  // public async show({limit, page, role, location, fitnessLevel, workoutType}:UserQuery): Promise<UserInfo[]> {
+  //   const queryParams = {
+  //     where: {
+  //       AND: {
+  //         role: role as UserRole,
+  //         fitnessLevel: fitnessLevel as FitnessLevel,
+  //         workoutType: undefined,
+  //         location: undefined,
+  //       }
+  //     },
+  //     take: limit,
+  //     skip: page > 0 ? limit * (page - 1) : undefined,
+  //   }
+  //   if (workoutType) {
+  //     queryParams.where.AND.workoutType = { hasSome: workoutType };
+  //   }
+  //   if (location) {
+  //     queryParams.where.AND.location = { in: location };
+  //   }
+  //   const users = await this.prisma.user.findMany(queryParams);
+  //   return users.map((user)=> adaptPrismaUser(user));
+  // }
 
   // public async update(id: number, item: UserInfoEntity): Promise<UserInfo> {
   //   const data ={
@@ -72,14 +76,14 @@ export class UserInfoRepository
   //       }
   //       delete data.passwordHash;
 
-  //       const newUser = await this.prisma.user.update({
-  //         where: { id },
+  //       const user = await this.prisma.user.update({
+  //         where: { userId:id },
   //         data
   //       });
   // return adaptPrismaUser(user);
   // }
 
   public async destroy(id: number): Promise<void> {
-    await this.prisma.user.delete({ where: { id } });
+    await this.prisma.user.delete({ where: { userId: id } });
   }
 }
